+++ s3client/types.go
@@ -8,11 +8,12 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go/service/s3/s3iface"
 )
 
 // S3Client - manages a persistent connection with downstream S3 bucket
 type S3Client struct {
-	s3Manager *s3.S3
+	s3Manager s3iface.S3API
 }
 
 // NewS3Client -  creates a new instance for S3Client with a aws session manager
@@ -48,4 +49,32 @@ type GetObjectOutput struct {
 	StorageClass    *string       `location:"header" locationName:"x-amz-storage-class" type:"string" enum:"StorageClass"`
 	TagCount        *int64        `location:"header" locationName:"x-amz-tagging-count" type:"integer"`
 	VersionId       *string       `location:"header" locationName:"x-amz-version-id" type:"string"`
+}
+
+// GetObject - talks to S3 to get content/byte-range from the bucket
+func (client *S3Client) GetObject(bucket, s3Path, byterange string) (*GetObjectOutput, error) {
+	getInput := &s3.GetObjectInput{
+		Bucket: aws.String(bucket),
+		Key:    aws.String(s3Path),
+		Range:  aws.String(byterange),
+	}
+
+	result, err := client.s3Manager.GetObject(getInput)
+	if err != nil {
+		return &GetObjectOutput{}, err
+	}
+
+	return &GetObjectOutput{
+		Body:            result.Body,
+		CacheControl:    result.CacheControl,
+		ContentEncoding: result.ContentEncoding,
+		ContentLength:   result.ContentLength,
+		ContentRange:    result.ContentRange,
+		ContentType:     result.ContentType,
+		ETag:            result.ETag,
+		Expiration:      result.Expiration,
+		LastModified:    result.LastModified,
+		StorageClass:    result.StorageClass,
+		TagCount:        result.TagCount,
+		VersionId:       result.VersionId}, nil
 }
